+++ jpa-criteria/impl/src/main/java/com/blazebit/persistence/criteria/impl/expression/function/SubstringFunction.java
@@ -19,7 +19,6 @@ package com.blazebit.persistence.criteria.impl.expression.function;
 import com.blazebit.persistence.criteria.impl.BlazeCriteriaBuilderImpl;
 import com.blazebit.persistence.criteria.impl.ParameterVisitor;
 import com.blazebit.persistence.criteria.impl.RenderContext;
-import com.blazebit.persistence.criteria.impl.expression.LiteralExpression;
 
 import javax.persistence.criteria.Expression;
 
@@ -46,14 +45,6 @@ public class SubstringFunction extends AbstractFunctionExpression<String> {
 
     public SubstringFunction(BlazeCriteriaBuilderImpl criteriaBuilder, Expression<String> value, Expression<Integer> start) {
         this(criteriaBuilder, value, start, (Expression<Integer>) null);
-    }
-
-    public SubstringFunction(BlazeCriteriaBuilderImpl criteriaBuilder, Expression<String> value, int start) {
-        this(criteriaBuilder, value, new LiteralExpression<Integer>(criteriaBuilder, start));
-    }
-
-    public SubstringFunction(BlazeCriteriaBuilderImpl criteriaBuilder, Expression<String> value, int start, int length) {
-        this(criteriaBuilder, value, new LiteralExpression<Integer>(criteriaBuilder, start), new LiteralExpression<Integer>(criteriaBuilder, length));
     }
 
     @Override